I must have been insane
Not too sure why I was making user media stream as a requirement, then
adapting that to also show on error.

To rectify that, local media stream is now an object - not required at
all, and the game will still play. :)

Also, GameBoard() now accep - ts an object parameter, not limited to
just a string - more control in the long run